requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pandas==2.1.4
dnspython==2.4.2
validate-email-address==1.0.0
Werkzeug==3.0.1